forcing inclusion of 0 SP values
Modified code to include 0 in list of Z and Y sensitivity parameters for
all cases except those in which a user-specified range is given and does
not include 0.  Modified plot function to substitute tau0 in for point
estimate in row/column corresponding to zeta = 0.  Done for both
treatSens and treatSens.BART